Refactored Configuration Files
